Added input checks

+++ main.py
@@ -13,13 +13,23 @@ def run():
     input = "./input/" + deck_name
     output = "./output/" + deck_name + ".csv"
 
+    # check if directory exists
+    if not os.path.exists(input):
+        print(f"⛔ Directory {input} does not exist.")
+        exit()
+
+    # check if output already exists
+    if os.path.exists(output):
+        print(f"⛔ File {output} already exists.")
+        exit()
+
     # get filenames from input directory
     paths = []
     for dirpath, _, filenames in os.walk(input):
         for filename in filenames:
             # get absolute path and remove spaces and commas
             path = os.path.abspath(os.path.join(dirpath, filename))
-            new_path = path.replace(" ", "").replace(",", "").replace("'", "").replace("ǵ", "")
+            new_path = path.replace(" ", "").replace(",", "").replace("'", "").replace("ǵ", "").replace("ń", "")
 
             # rename files and append paths
             os.rename(path, new_path)
@@ -29,10 +39,10 @@ def run():
     with open(output, 'w') as f:
         f.write("@image" + "\n")
         for path in paths:
-            # print(path)
+            # print(f"Path: {path}")
             f.write(path + "\n")
 
-    print(f"✅ PDF created successfully at {output}")
+    print(f"✅ CSV file created successfully at {output}")
     
 
 if __name__ == "__main__":
